Notifcations On Post to everypage
Notifications should work on every page now

+++ Pages/Dashboard.cshtml.cs
@@ -87,10 +87,21 @@ namespace CS3750_PlanetExpressLMS.Pages
             return Page();
         }
 
-        public async Task<IActionResult> OnPost(int id)
+        public async Task<IActionResult> OnPostClearNotification(int id)
         {
+            // Access the current session
+            PlanetExpressSession session = new PlanetExpressSession(HttpContext);
+
+            // Make sure a user is logged in
+            user = session.GetUser();
+
+            if (user == null)
+            {
+                return RedirectToPage("Login");
+            }
+
             notificationRepository.Delete(id);
-            return RedirectToPage("/Dashboard");
+            return RedirectToPage("Dashboard");
         }
     }
 }
